AuthorizationService/Service/requirements.txt
flask
flask_jwt_extended
dotenv
cryptography
mysql-connector-python